requirements.txt
python-telegram-bot==20.0
pymongo
python-dotenv
requests
pillow
google-generativeai
